- Fixed calls with I8/U8 return values (passes CallI8 tests) - Fixed bug in IL ldobj/ldind instruction handling, fixes LdlocaU8_CheckValue tests - Moved expansion of return statement to platform specific transformation stage - 55 failures left (Passes 2004 of 2059 tests)
git-svn-id: svn://localhost/mosa@225 45b04e34-05f1-4d87-a130-40a8491d5c8b

+++ Test/Mosa/Runtime/CompilerFramework/IL/Call.cs
@@ -155,7 +155,7 @@ namespace Test.Mosa.Runtime.CompilerFramework.IL
         /// </summary>
         /// <param name="arg"></param>
         /// <returns></returns>
-        delegate bool B__I8(int arg);
+        delegate bool B__I8(long arg);
         /// <summary>
         /// 
         /// </summary>
